cpct_memset_f64 bad pointer type corrected

+++ memutils/cpct_memutils.c
@@ -44,7 +44,7 @@ void cpct_memcpy(void* to, const void* from, u16 size)
 
 void cpct_memset_f64(void *array, u16 value, u16 size)
 {
-	u16* data = (u16*)GetMemory(array);
+	u8* data = (u8*)GetMemory(array);
 
 	for (int i = 0; i < size; i++)
 		data[i] = value;
